refactor/#321 : DTO 클래스 dto패키지로 이동
