+++ BE/exceed/src/main/java/com/gaebaljip/exceed/auth/adapter/AuthController.java
@@ -16,7 +16,7 @@ import com.gaebaljip.exceed.common.ApiResponse;
 import com.gaebaljip.exceed.common.ApiResponseGenerator;
 import com.gaebaljip.exceed.common.swagger.ApiErrorExceptionsExample;
 import com.gaebaljip.exceed.dto.request.LoginRequest;
-import com.gaebaljip.exceed.dto.response.LoginResponseDTO;
+import com.gaebaljip.exceed.dto.LoginResponseDTO;
 import com.gaebaljip.exceed.security.AuthConstants;
 
 import io.swagger.v3.oas.annotations.Operation;
