+++ BE/exceed/src/main/java/com/gaebaljip/exceed/auth/application/port/AuthService.java
@@ -7,7 +7,7 @@ import org.springframework.transaction.annotation.Transactional;
 import com.gaebaljip.exceed.auth.application.port.in.AuthUsecase;
 import com.gaebaljip.exceed.auth.exception.PasswordMismatchException;
 import com.gaebaljip.exceed.dto.request.LoginRequest;
-import com.gaebaljip.exceed.dto.response.LoginResponseDTO;
+import com.gaebaljip.exceed.dto.LoginResponseDTO;
 import com.gaebaljip.exceed.member.adapter.out.persistence.MemberEntity;
 import com.gaebaljip.exceed.member.application.port.out.MemberPort;
 import com.gaebaljip.exceed.security.domain.JwtManager;
