+++ BE/exceed/src/main/java/com/gaebaljip/exceed/auth/application/port/in/AuthUsecase.java
@@ -3,7 +3,7 @@ package com.gaebaljip.exceed.auth.application.port.in;
 import org.springframework.stereotype.Component;
 
 import com.gaebaljip.exceed.dto.request.LoginRequest;
-import com.gaebaljip.exceed.dto.response.LoginResponseDTO;
+import com.gaebaljip.exceed.dto.LoginResponseDTO;
 
 @Component
 public interface AuthUsecase {
