+++ BE/exceed/src/main/java/com/gaebaljip/exceed/dto/EatMealFoodDTO.java
@@ -1,4 +1,4 @@
-package com.gaebaljip.exceed.dto.request;
+package com.gaebaljip.exceed.dto;
 
 import lombok.Builder;
