+++ BE/exceed/src/main/java/com/gaebaljip/exceed/dto/GetPageableFoodDTO.java
@@ -1,3 +1,3 @@
-package com.gaebaljip.exceed.dto.response;
+package com.gaebaljip.exceed.dto;
 
 public record GetPageableFoodDTO(Long id, String name) {}
