+++ BE/exceed/src/main/java/com/gaebaljip/exceed/dto/LoginResponseDTO.java
@@ -1,4 +1,4 @@
-package com.gaebaljip.exceed.dto.response;
+package com.gaebaljip.exceed.dto;
 
 import lombok.Builder;
