+++ BE/exceed/src/main/java/com/gaebaljip/exceed/dto/MealRecordDTO.java
@@ -1,8 +1,9 @@
-package com.gaebaljip.exceed.dto.response;
+package com.gaebaljip.exceed.dto;
 
 import java.time.LocalTime;
 import java.util.List;
 
+import com.gaebaljip.exceed.dto.FoodDTO;
 import com.gaebaljip.exceed.meal.domain.MealType;
 
 import lombok.Builder;
