+++ BE/exceed/src/main/java/com/gaebaljip/exceed/dto/request/EatMealRequest.java
@@ -3,6 +3,7 @@ package com.gaebaljip.exceed.dto.request;
 import java.util.List;
 
 import com.gaebaljip.exceed.common.annotation.Enum;
+import com.gaebaljip.exceed.dto.EatMealFoodDTO;
 import com.gaebaljip.exceed.meal.domain.MealType;
 
 import lombok.Builder;
