+++ BE/exceed/src/main/java/com/gaebaljip/exceed/dto/response/GetAnalysisResponse.java
@@ -1,5 +1,7 @@
 package com.gaebaljip.exceed.dto.response;
 
+import com.gaebaljip.exceed.dto.AnalysisDTO;
+
 import java.util.List;
 
 public record GetAnalysisResponse(List<AnalysisDTO> getAnalyses) {}
