+++ BE/exceed/src/main/java/com/gaebaljip/exceed/dto/response/GetMealFoodResponse.java
@@ -2,6 +2,7 @@ package com.gaebaljip.exceed.dto.response;
 
 import java.util.List;
 
+import com.gaebaljip.exceed.dto.MealRecordDTO;
 import lombok.Builder;
 
 public record GetMealFoodResponse(
