+++ BE/exceed/src/main/java/com/gaebaljip/exceed/dto/response/GetMealResponse.java
@@ -1,5 +1,9 @@
 package com.gaebaljip.exceed.dto.response;
 
+import com.gaebaljip.exceed.dto.CurrentMealDTO;
+import com.gaebaljip.exceed.dto.MaintainMealDTO;
+import com.gaebaljip.exceed.dto.TargetMealDTO;
+
 public record GetMealResponse(
         MaintainMealDTO maintainMealDTO,
         TargetMealDTO targetMealDTO,
