+++ BE/exceed/src/main/java/com/gaebaljip/exceed/meal/application/EatMealService.java
@@ -8,7 +8,7 @@ import javax.transaction.Transactional;
 
 import org.springframework.stereotype.Service;
 
-import com.gaebaljip.exceed.dto.request.EatMealFoodDTO;
+import com.gaebaljip.exceed.dto.EatMealFoodDTO;
 import com.gaebaljip.exceed.food.adapter.out.FoodEntity;
 import com.gaebaljip.exceed.food.application.port.out.FoodPort;
 import com.gaebaljip.exceed.meal.adapter.out.MealEntity;
