+++ BE/exceed/src/main/java/com/gaebaljip/exceed/meal/application/GetSpecificMealService.java
@@ -7,11 +7,11 @@ import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.gaebaljip.exceed.dto.request.TodayMealDTO;
-import com.gaebaljip.exceed.dto.response.CurrentMealDTO;
-import com.gaebaljip.exceed.dto.response.FoodDTO;
-import com.gaebaljip.exceed.dto.response.MealRecordDTO;
-import com.gaebaljip.exceed.dto.response.SpecificMealDTO;
+import com.gaebaljip.exceed.dto.TodayMealDTO;
+import com.gaebaljip.exceed.dto.CurrentMealDTO;
+import com.gaebaljip.exceed.dto.FoodDTO;
+import com.gaebaljip.exceed.dto.MealRecordDTO;
+import com.gaebaljip.exceed.dto.SpecificMealDTO;
 import com.gaebaljip.exceed.meal.application.port.in.GetSpecificMealQuery;
 import com.gaebaljip.exceed.meal.application.port.out.DailyMealPort;
 import com.gaebaljip.exceed.meal.application.port.out.PresignedUrlPort;
