+++ BE/exceed/src/main/java/com/gaebaljip/exceed/meal/application/port/in/GetCurrentMealQuery.java
@@ -2,7 +2,7 @@ package com.gaebaljip.exceed.meal.application.port.in;
 
 import org.springframework.stereotype.Component;
 
-import com.gaebaljip.exceed.dto.response.CurrentMealDTO;
+import com.gaebaljip.exceed.dto.CurrentMealDTO;
 
 @Component
 public interface GetCurrentMealQuery {
