+++ BE/exceed/src/main/java/com/gaebaljip/exceed/meal/application/port/in/GetSpecificMealQuery.java
@@ -4,7 +4,7 @@ import java.time.LocalDateTime;
 
 import org.springframework.stereotype.Component;
 
-import com.gaebaljip.exceed.dto.response.SpecificMealDTO;
+import com.gaebaljip.exceed.dto.SpecificMealDTO;
 
 @Component
 public interface GetSpecificMealQuery {
