+++ BE/exceed/src/main/java/com/gaebaljip/exceed/member/application/UpdateWeightService.java
@@ -3,7 +3,7 @@ package com.gaebaljip.exceed.member.application;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.gaebaljip.exceed.dto.UpdateWeightResponse;
+import com.gaebaljip.exceed.dto.response.UpdateWeightResponse;
 import com.gaebaljip.exceed.member.adapter.out.persistence.HistoryEntity;
 import com.gaebaljip.exceed.member.adapter.out.persistence.MemberEntity;
 import com.gaebaljip.exceed.member.application.port.in.UpdateWeightCommand;
