+++ BE/exceed/src/main/java/com/gaebaljip/exceed/member/application/port/in/GetMaintainMealUsecase.java
@@ -4,7 +4,7 @@ import java.time.LocalDateTime;
 
 import org.springframework.stereotype.Component;
 
-import com.gaebaljip.exceed.dto.response.MaintainMealDTO;
+import com.gaebaljip.exceed.dto.MaintainMealDTO;
 
 @Component
 public interface GetMaintainMealUsecase {
