+++ BE/exceed/src/main/java/com/gaebaljip/exceed/member/application/port/in/GetTargetMealUsecase.java
@@ -4,7 +4,7 @@ import java.time.LocalDateTime;
 
 import org.springframework.stereotype.Component;
 
-import com.gaebaljip.exceed.dto.response.TargetMealDTO;
+import com.gaebaljip.exceed.dto.TargetMealDTO;
 
 @Component
 public interface GetTargetMealUsecase {
