+++ BE/exceed/src/main/java/com/gaebaljip/exceed/member/application/port/in/UpdateWeightUsecase.java
@@ -2,7 +2,7 @@ package com.gaebaljip.exceed.member.application.port.in;
 
 import org.springframework.stereotype.Component;
 
-import com.gaebaljip.exceed.dto.UpdateWeightResponse;
+import com.gaebaljip.exceed.dto.response.UpdateWeightResponse;
 
 @Component
 public interface UpdateWeightUsecase {
