+++ BE/exceed/src/main/java/com/gaebaljip/exceed/nutritionist/application/GetAnalysisService.java
@@ -11,8 +11,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.gaebaljip.exceed.common.annotation.Timer;
 import com.gaebaljip.exceed.dto.request.GetAnalysisRequest;
-import com.gaebaljip.exceed.dto.request.MonthlyMealDTO;
-import com.gaebaljip.exceed.dto.response.AnalysisDTO;
+import com.gaebaljip.exceed.dto.MonthlyMealDTO;
+import com.gaebaljip.exceed.dto.AnalysisDTO;
 import com.gaebaljip.exceed.dto.response.GetAnalysisResponse;
 import com.gaebaljip.exceed.meal.domain.DailyMeal;
 import com.gaebaljip.exceed.meal.domain.Meal;
