+++ BE/exceed/src/test/java/com/gaebaljip/exceed/meal/adapter/EatMealIntegrationTest.java
@@ -16,7 +16,7 @@ import org.springframework.test.web.servlet.ResultActions;
 
 import com.gaebaljip.exceed.common.IntegrationTest;
 import com.gaebaljip.exceed.common.WithMockUser;
-import com.gaebaljip.exceed.dto.request.EatMealFoodDTO;
+import com.gaebaljip.exceed.dto.EatMealFoodDTO;
 import com.gaebaljip.exceed.dto.request.EatMealRequest;
 import com.gaebaljip.exceed.meal.adapter.out.MealRepository;
 import com.gaebaljip.exceed.meal.application.port.out.PresignedUrlPort;
